commit 49f49d0
Author: Peter Kwekkeboom <peter.kwekkeboom@dekimo.com>
Date:   Tue Dec 12 11:51:35 2023 +0100

    change: use new write log function and also save EOS battery state.

commit d40b554
Author: Peter Kwekkeboom <peter.kwekkeboom@dekimo.com>
Date:   Tue Dec 12 10:25:06 2023 +0100

    change: state of health is uint8, nog uint16.

commit 1e289c7
Author: Peter Kwekkeboom <peter.kwekkeboom@dekimo.com>
Date:   Tue Dec 12 10:14:31 2023 +0100

    change: add new log function to dataflash.

commit e2bfc75
Author: Peter Kwekkeboom <peter.kwekkeboom@dekimo.com>
Date:   Tue Dec 12 10:13:57 2023 +0100

    change: rename old writeLog

commit c3178cd
Author: Peter Kwekkeboom <peter.kwekkeboom@dekimo.com>
Date:   Tue Dec 12 10:11:35 2023 +0100

    change: rewrite with struct in dataflash logdata array.

commit 79eeaee
Author: Peter Kwekkeboom <peter.kwekkeboom@dekimo.com>
Date:   Tue Dec 12 09:25:08 2023 +0100

    change: rewrite mainTask for save sensorModule data to dataflash.

commit 2320762
Author: Peter Kwekkeboom <peter.kwekkeboom@dekimo.com>
Date:   Tue Dec 12 09:21:27 2023 +0100

    change: add definition for MFM sensor data and MFM main data for logging.

+++ App/BatMon_BQ35100/bq35100.c
@@ -130,9 +130,9 @@ uint16_t bq35100_getMeasuredZ(void)
  *
  * @return The state of health in a percentage.
  */
-uint16_t bq35100_getStateOfHealth(void)
+uint8_t bq35100_getStateOfHealth(void)
 {
-  uint16_t stateOfHealth;
+  uint8_t stateOfHealth;
   uint8_t data[2];
 
   HAL_I2C_Mem_Read(bq35100Handle, BQ35100_ADDRESS, REG_STATE_OF_HEALTH, 1, data, 2, BQ35100_I2C_WAIT);
